requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
google-generativeai==0.3.2
